ダミーニュース処理をダミーサービス内に限定
