+++ WondayWall.Android/app/src/main/kotlin/com/studiofreesia/wondaywall/ui/screens/data/DataViewModel.kt
@@ -73,8 +73,7 @@ class DataViewModel(
 
     // ニュースを取得する
     private suspend fun loadNews(rssSources: List<String>) {
-        val debugConfig = appConfigService.getDebugConfig()
-        if (rssSources.isEmpty() && !debugConfig.useDummyAiService) return
+        if (rssSources.isEmpty()) return
         _uiState.value = _uiState.value.copy(isLoadingNews = true)
         try {
             val result = contextService.buildPromptContext()
